Add user-filter
The `UserFilterComponent` is a way to allow users to select target values for features of users they would like to put in a collection. Much of the code here came from `UserListComponent`, which used to include the filtering right in the same view.

I found the previous structure a little confusing because the template took care of all of the filtering and displaying the collection of users that resulted from the filtering. I thought it might help to separate the filtering from the display of the collection of users.

It was exciting to see how the e2e tests were not impacted by this change.