Add schemas to `User` and `UserRole`
This also pulls `UserRole` out into its own file.

The new `validateUser` function can be used to determine whether any JavaScript value is in fact an instance of the type `User`.

I had to make `_id` optional, since there are numerous places where we don't (yet) have an ID. As an example, when adding a user, we don't have an ID because that will come from the server when we actually add it.