Validate JSON received by `getUserById` is a `User`
The key part of this is the change to `getUserById`, where we now use the `validateUser()` function to check that the `user` value we got in the HTTP response from the server really is an instance of `User`. If it's not, we throw an error.

The other changes to this file are introductions of `!` to say that code should fail is a value is `null`. For example:

```typescript
return this.httpClient!.get<User[]>(...
```

the `!` says that if `this.httpClient` is `null` this should fail straightaway. This was only necessary because I had to say that the injected `httpClient` in the constructor could have the value `null`. I'm not happy about adding `null` as an option there, but I can figure out how to get the injection to work so that the mocked version of the service doesn't have to provide an `HttpClient`.

I _think_ the solution to that is to get rid of the `MockUserService` and instead just stub what we need where we need it. Then we could get rid of `HttpClient | null`, and we wouldn't need all those `!`s.