Get first two tests with validation to pass
I think this demonstrates how to fix the validation code in Javalin v6. While I was here I also expanded fair bit on the commenting for these two tests.

I need to continue to make similar changes on the remaining tests that include validation, all of which are currently commented out (so I can compile and run the tests).