Get failed validation tests to pass
This fixes the tests that check that validation fails as expected on illegal input (in this case a string that can't be parsed to an integer).

I think I figured out how to get the details out of a `ValidationException` as well, so I added checks for more of the underlying cause of the failure. In doing that, it seemed reasonable to expect (and include in the test) that the validation error messages would include the offending string. That required adding a mock for `ctx.queryParam(AGE_KEY)` in the tests so we could call that in the controller to get the string that was passed in.